Move file collection snapshotting to core
