+++ subprojects/core/src/main/java/org/gradle/internal/fingerprint/FileCollectionSnapshotter.java
@@ -1,0 +1,32 @@
+/*
+ * Copyright 2019 the original author or authors.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.gradle.internal.fingerprint;
+
+import org.gradle.api.file.FileCollection;
+import org.gradle.internal.snapshot.FileSystemSnapshot;
+
+import java.util.List;
+
+/**
+ * Service for snapshotting {@link FileCollection}s.
+ */
+public interface FileCollectionSnapshotter {
+    /**
+     * Returns snapshots of the roots of a file collection.
+     */
+    List<FileSystemSnapshot> snapshot(FileCollection fileCollection);
+}
