+++ subprojects/core/src/main/java/org/gradle/internal/fingerprint/classpath/impl/DefaultCompileClasspathFingerprinter.java
@@ -22,17 +22,17 @@ import org.gradle.api.internal.changedetection.state.CachingResourceHasher;
 import org.gradle.api.internal.changedetection.state.ResourceSnapshotterCacheService;
 import org.gradle.api.tasks.CompileClasspathNormalizer;
 import org.gradle.api.tasks.FileNormalizer;
+import org.gradle.internal.fingerprint.FileCollectionSnapshotter;
 import org.gradle.internal.fingerprint.classpath.CompileClasspathFingerprinter;
 import org.gradle.internal.fingerprint.impl.AbstractFileCollectionFingerprinter;
-import org.gradle.internal.snapshot.FileSystemSnapshotter;
 
 public class DefaultCompileClasspathFingerprinter extends AbstractFileCollectionFingerprinter implements CompileClasspathFingerprinter {
-    public DefaultCompileClasspathFingerprinter(ResourceSnapshotterCacheService cacheService, FileSystemSnapshotter fileSystemSnapshotter, StringInterner stringInterner) {
+    public DefaultCompileClasspathFingerprinter(ResourceSnapshotterCacheService cacheService, FileCollectionSnapshotter fileCollectionSnapshotter, StringInterner stringInterner) {
         super(ClasspathFingerprintingStrategy.compileClasspath(
                     new CachingResourceHasher(new AbiExtractingClasspathResourceHasher(), cacheService),
                     cacheService,
                     stringInterner
-                ), fileSystemSnapshotter);
+                ), fileCollectionSnapshotter);
     }
 
     @Override
