+++ subprojects/core/src/main/java/org/gradle/internal/fingerprint/impl/RelativePathFileCollectionFingerprinter.java
@@ -18,13 +18,13 @@ package org.gradle.internal.fingerprint.impl;
 
 import org.gradle.api.internal.cache.StringInterner;
 import org.gradle.api.tasks.FileNormalizer;
+import org.gradle.internal.fingerprint.FileCollectionSnapshotter;
 import org.gradle.internal.fingerprint.RelativePathInputNormalizer;
-import org.gradle.internal.snapshot.FileSystemSnapshotter;
 
 public class RelativePathFileCollectionFingerprinter extends AbstractFileCollectionFingerprinter {
 
-    public RelativePathFileCollectionFingerprinter(StringInterner stringInterner, FileSystemSnapshotter fileSystemSnapshotter) {
-        super(new RelativePathFingerprintingStrategy(stringInterner), fileSystemSnapshotter);
+    public RelativePathFileCollectionFingerprinter(StringInterner stringInterner, FileCollectionSnapshotter fileCollectionSnapshotter) {
+        super(new RelativePathFingerprintingStrategy(stringInterner), fileCollectionSnapshotter);
     }
 
     @Override
